backend/requirements.txt
fastapi[standard]
sqlalchemy
uvicorn[standard]
pydantic
python-dotenv
pydantic-settings
alembic
psycopg[binary]
sqlalchemy
authlib
python-jose
httpx
psycopg2-binary   
itsdangerous
requests
python-jose